BackgroundFetcher: Explicitly cast the current tile number to a signed type when handling the second tile data method

+++ src/ppu/fifo/BackgroundFetcher.cpp
@@ -69,7 +69,8 @@ void gbtest::BackgroundFetcher::executeState()
                             + m_ppuRegisters.lcdPositionAndScrolling.yLcdCoordinate) % 8);
         }
         else {
-            m_currentTileData = m_vram.getVramTileData().getTileLineUsingSecondMethod(m_currentTileNumber,
+            m_currentTileData = m_vram.getVramTileData().getTileLineUsingSecondMethod(
+                    static_cast<int8_t>(m_currentTileNumber),
                     (m_ppuRegisters.lcdPositionAndScrolling.yScroll
                             + m_ppuRegisters.lcdPositionAndScrolling.yLcdCoordinate) % 8);
         }
